Extract connection projection to separate files
