+++ packages/graphql/src/translate/cypher-builder/expressions/procedures/apoc/date.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import type { CypherEnvironment } from "src/translate/cypher-builder/Environment";
+import type { CypherEnvironment } from "../../../Environment";
 import { CypherASTNode } from "../../../CypherASTNode";
 import type { PropertyRef, Variable } from "../../../CypherBuilder";
